This is the first pass at building the ANS table.
This is likely not going to be used in the final decoding step. It
turns out that OpenCL cannot synchronize memory accesses across
work-groups. This means that we can't build the rANS table on the
GPU itself and we have to build it in software. The drawbacks are
twofold:

1. We have to upload the table, which is much larger than the
alphabet or anything else that we upload except maybe the compressed
rANS data.

2. We have to spend time building it sequentially on the CPU.

This discouraging development makes me think that perhaps we should
just entirely switch over to CUDA, although the restriction to NVIDIA
devices makes that approach unappealing as well.