For the ANS OpenCL kernel, read all data from a single buffer
This eliminates the need to do multiple host-side allocations, and we can
actually write our encoder to set up the data in a way that the ocl decoder
is expecting it. In this way we can avoid a couple of allocations and it makes
sending multiple chunks at once easier.