Use different OpenCL device if debugging on Intel CPU
In particular, the OpenCL CPU runtime on intel chips using visual
studio comes with a pretty good debugger, so we should take advantage
of it. If we detect something along the lines of an Intel CPU in debug
mode, then try to use it in debug mode... This is super not cross platform
but I hope I put enough guards so that it doesn't trigger unless you
support it. The default is still to ask for a GPU device.